Add files via upload
This is a fertilizer recommendation system with an accuracy of 100%. It involves importing the required libraries, analyzing the shape of the dataset, and dealing with the null or missing values. We then plot a graph for correlation of the data and visualize the distribution of each of its features. Encoding is applied to map each fertilizer name with its respective number and make recommendation of the suitable fertilizer using the input values.